Some bugfixes

See merge request cmdevs/ColonialMarines!3659

+++ code/modules/mob/living/carbon/xenomorph/attack_alien.dm
@@ -682,7 +682,7 @@
 //Xenomorphs can't use machinery, not even the "intelligent" ones
 //Exception is Queen and shuttles, because plot power
 /obj/structure/machinery/attack_alien(mob/living/carbon/Xenomorph/M)
-	if(health <= 0 || unacidable)
+	if(unslashable || health <= 0)
 		to_chat(M, SPAN_WARNING("You stare at \the [src] cluelessly."))
 	else
 		M.animation_attack_on(src)
@@ -691,7 +691,8 @@
 		if(health <= 0)
 			M.visible_message(SPAN_DANGER("\The [M] slices [src] apart!"), \
 			SPAN_DANGER("You slice [src] apart!"), null, 5)
-			qdel(src)
+			if (!unacidable)
+				qdel(src)
 		else
 			M.visible_message(SPAN_DANGER("[M] slashes [src]!"), \
 			SPAN_DANGER("You slash [src]!"), null, 5)
